--- src/cascadia/TerminalConnection/ConhostConnection.idl
@@ -1,15 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-import "ITerminalConnection.idl";
-
-namespace Microsoft.Terminal.TerminalConnection
-{
-    [default_interface] runtimeclass ConhostConnection : ITerminalConnection
-    {
-        ConhostConnection(String cmdline, String startingDirectory, String startingTitle, UInt32 rows, UInt32 columns, Guid guid);
-
-        Guid Guid { get; };
-    };
-
-}
