+++ src/cascadia/TerminalConnection/ConptyConnection.idl
@@ -8,7 +8,8 @@ namespace Microsoft.Terminal.TerminalConnection
     [default_interface]
     runtimeclass ConptyConnection : ITerminalConnection
     {
-        ConptyConnection(String commandline, UInt32 initialRows, UInt32 initialCols);
+        ConptyConnection(String cmdline, String startingDirectory, String startingTitle, UInt32 rows, UInt32 columns, Guid guid);
+        Guid Guid { get; };
     };
 
 }
