+++ src/cascadia/TerminalConnection/TerminalConnection.vcxproj
@@ -35,8 +35,8 @@
     <ClInclude Include="AzureConnection-ARM64.h" Condition="'$(Platform)'=='ARM64'" />
     <ClInclude Include="AzureConnectionStrings.h" Condition="'$(Platform)'!='ARM64'" />
     <ClInclude Include="pch.h" />
-    <ClInclude Include="ConhostConnection.h">
-      <DependentUpon>ConhostConnection.idl</DependentUpon>
+    <ClInclude Include="ConptyConnection.h">
+      <DependentUpon>ConptyConnection.idl</DependentUpon>
     </ClInclude>
     <ClInclude Include="EchoConnection.h">
       <DependentUpon>EchoConnection.idl</DependentUpon>
@@ -49,17 +49,17 @@
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader>Create</PrecompiledHeader>
     </ClCompile>
-    <ClCompile Include="ConhostConnection.cpp">
-      <DependentUpon>ConhostConnection.idl</DependentUpon>
-    </ClCompile>
     <ClCompile Include="EchoConnection.cpp">
       <DependentUpon>EchoConnection.idl</DependentUpon>
+    </ClCompile>
+    <ClCompile Include="ConptyConnection.cpp">
+      <DependentUpon>ConptyConnection.idl</DependentUpon>
     </ClCompile>
     <ClCompile Include="$(GeneratedFilesDir)module.g.cpp" />
   </ItemGroup>
   <ItemGroup>
     <Midl Include="ITerminalConnection.idl" />
-    <Midl Include="ConhostConnection.idl" />
+    <Midl Include="ConptyConnection.idl" />
     <Midl Include="EchoConnection.idl" />
     <Midl Include="AzureConnection.idl" />
   </ItemGroup>
@@ -77,6 +77,9 @@
     <ProjectReference Include="$(OpenConsoleDir)src\types\lib\types.vcxproj">
       <Project>{18D09A24-8240-42D6-8CB6-236EEE820263}</Project>
     </ProjectReference>
+    <ProjectReference Include="..\..\winconpty\winconpty.vcxproj">
+      <Project>{58a03bb2-df5a-4b66-91a0-7ef3ba01269a}</Project>
+    </ProjectReference>
   </ItemGroup>
 
   <Import Project="$(OpenConsoleDir)src\cppwinrt.build.post.props" />
@@ -88,4 +91,10 @@
     </PropertyGroup>
     <Error Condition="!Exists('..\..\..\packages\vcpkg-cpprestsdk.2.10.0\build\native\vcpkg-cpprestsdk.targets')" Text="$([System.String]::Format('$(ErrorText)', '..\..\..\packages\vcpkg-cpprestsdk.2.10.0\build\native\vcpkg-cpprestsdk.targets'))" />
   </Target>
+
+  <ItemDefinitionGroup>
+    <Link>
+      <AdditionalDependencies>$(OpenConsoleCommonOutDir)\conpty.lib;%(AdditionalDependencies)</AdditionalDependencies>
+    </Link>
+  </ItemDefinitionGroup>
 </Project>
